+++ .gitignore
@@ -1,9 +1,5 @@
 # dependencies
 /node_modules
-/functions/node_modules
-
-# production
-/build
 
 # misc
 .DS_Store
