--- firebase.json
@@ -1,15 +1,0 @@
-{
-	"functions": {
-		"source": "functions"
-	},
-	"hosting": {
-		"public": "build",
-		"ignore": ["firebase.json", "**/.*", "**/node_modules/**"],
-		"rewrites": [
-			{
-				"source": "/api/**",
-				"function": "api"
-			}
-		]
-	}
-}
