+++ package.json
@@ -1,54 +1,19 @@
 {
-	"name": "corona-mask-market",
-	"version": "0.2.0",
-	"private": true,
+	"name": "covid19-kr-server",
+	"scripts": {
+		"start": "nodemon index.js",
+		"dev": "concurrently \"npm run dev:server\" \"npm run dev:client\"",
+		"dev:server": "npm start",
+		"dev:client": "cd client && npm start",
+		"heroku-postbuild": "cd client && npm install && npm run build"
+	},
 	"dependencies": {
-		"@testing-library/jest-dom": "^4.2.4",
-		"@testing-library/react": "^9.3.2",
-		"@testing-library/user-event": "^7.1.2",
-		"@toast-ui/react-chart": "^1.0.2",
 		"axios": "^0.21.1",
-		"concurrently": "^6.0.1",
+		"concurrently": "^6.0.2",
 		"cors": "^2.8.5",
 		"dotenv": "^8.2.0",
 		"express": "^4.17.1",
-		"firebase": "^7.14.2",
-		"http-proxy-middleware": "^1.1.0",
-		"immer": "^6.0.2",
-		"node-sass": "^4.13.1",
-		"nodemon": "^2.0.7",
-		"react": "^16.13.0",
-		"react-dom": "^16.13.0",
-		"react-icons": "^3.9.0",
-		"react-router-dom": "^5.1.2",
-		"react-scripts": "^4.0.3",
-		"react-spring": "^8.0.27",
-		"styled-components": "^5.0.1",
-		"tui-chart": "^3.11.2"
+		"nodemon": "^2.0.7"
 	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"deploy": "gh-pages -d build"
-	},
-	"eslintConfig": {
-		"extends": "react-app"
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"gh-pages": "^2.2.0"
-	}
+	"private": true
 }
