--- public/index.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<meta name="description" content="코로나 마스크 판매 사이트 모아보기" />
-		<meta name="author" content="Subin Ahn" />
-		<meta property="og:type" content="website" />
-		<meta property="og:title" content="코로나 마스크 마켓" />
-		<meta
-			property="og:description"
-			content="코로나 마스크 판매 사이트 모아보기"
-		/>
-		<meta
-			property="og:image"
-			content="https://image.flaticon.com/icons/png/128/1033/1033165.png"
-		/>
-		<meta property="og:url" content="#" />
-		<meta
-			name="keywords"
-			content="코로나, 코로나 마스크, 착한 마스크, 정직한 마스크, 마스크, kf94, kf80, n95, kf94 마스크, kf80 마스크, n95 마스크, 일회용 마스크, 마스크 판매 사이트 모음, 마스크 판매 사이트"
-		/>
-
-		<link rel="icon" href="favicon.svg" />
-
-		<link rel="apple-touch-icon" href="favicon.svg" />
-		<link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-
-		<!-- <script>
-    // http -> https
-    // if (window.location.protocol === "http:" && window.location.hostname !== "localhost") {
-    //   location.replace("https://getcoronamask.kr")
-    // }
-  </script> -->
-
-		<script
-			type="text/javascript"
-			src="//dapi.kakao.com/v2/maps/sdk.js?appkey=%REACT_APP_KAKAOMAP_API_KEY%"
-		></script>
-
-		<title>코로나 마스크 마켓</title>
-	</head>
-
-	<body>
-		<noscript>You need to enable JavaScript to run this app.</noscript>
-		<div id="root"></div>
-
-		<!-- The core Firebase JS SDK is always required and must be listed first -->
-		<script src="https://www.gstatic.com/firebasejs/7.13.1/firebase-app.js"></script>
-
-		<!-- TODO: Add SDKs for Firebase products that you want to use
-      https://firebase.google.com/docs/web/setup#available-libraries -->
-		<script src="https://www.gstatic.com/firebasejs/7.13.1/firebase-analytics.js"></script>
-	</body>
-</html>
