--- src/components/CheerKR.js
@@ -1,23 +1,0 @@
-import React from 'react';
-
-function CheerKR() {
-  return (
-    <>
-        <section>
-            <div className="head-section">
-              <h1>
-                #서포트
-              </h1>
-              <h3 className="head-section-text">
-                응원의 메세지를 남겨보세요!
-              </h3>
-            </div>
-        </section>
-
-        <div className="waiting">
-          이 페이지는 준비중입니다 :)
-        </div>
-    </>
-  );
-}
-export default CheerKR;
