--- src/masks.json
@@ -1,43 +1,0 @@
-[{
-    "image": "url(\"http://trencadism.com/web/product/big/20200313/3a30011adce080ab6e3bacf1a1153e1e.jpg\")",
-    "price": "16,500원",
-    "name": "[트렌카디즘] KF94마스크 브랜드 랜덤 대형 (10매)",
-    "time": "18:00",
-    "link": "https://trencadism.com/product/detail.html?product_no=1187"
-}, {
-    "image": "url(\"https://shop-phinf.pstatic.net/20200229_263/1582911182588lW4kX_JPEG/20273515981245379_708374544.jpg?type=m510\")",
-    "price": "14,900원",
-    "name": "[미마몰] 미마마스크 어린이 미세먼지 황사 보건용마스크 10개입(KF80)",
-    "time": "09:00",
-    "link": "https://smartstore.naver.com/aseado/products/4837266971"
-}, {
-    "image": "url(\"https://shop-phinf.pstatic.net/20200229_36/1582908044714SGwOl_JPEG/20269587155943437_1452556976.jpg?type=m510\")",
-    "price": "14,900원",
-    "name": "[미마몰] 미마마스크 미마 미세먼지 황사 보건용마스크 10개입(KF94)",
-    "time": "09:00",
-    "link": "https://smartstore.naver.com/aseado/products/4837257765"
-}, {
-    "image": "url(\"https://shop-phinf.pstatic.net/20200318_72/1584523139645CdLxo_JPEG/21885473202923997_1414107817.jpg?type=m510\")",
-    "price": "11,300원",
-    "name": "[아에르] KF94 아에르 스탠다드 베이직 마스크 화이트 S, M / 블랙 L(10매)",
-    "time": "09:45",
-    "link": "https://smartstore.naver.com/aer-shop/products/4722827602"
-}, {
-    "image": "url(\"https://shop-phinf.pstatic.net/20200309_23/1583714716310Qd6jx_PNG/21074550857298556_1589413599.png?type=m510\")",
-    "price": "4,900원",
-    "name": "[뉴네퓨어] 뉴네퓨어 식약처허가 KF94 미세먼지 끈길이조절 황사 마스크 화이트 5매",
-    "time": "10:00",
-    "link": "https://smartstore.naver.com/gonggami/products/4705579501"
-}, {
-    "image": "url(\"https://image.msscdn.net/images/goods_img/20191231/1260411/1260411_2_500.jpg\")",
-    "price": "29,900원",
-    "name": "[무신사] 스탠다드 밸브 마스크 KF 94 20팩 [블랙]",
-    "time": "12:00",
-    "link": "https://store.musinsa.com/app/raffle/views/65"
-}, {
-    "image": "url(\"https://shop-phinf.pstatic.net/20200212_124/1581494656106CzrYU_JPEG/18856198683634047_165341866.jpg?type=m510\")",
-    "price": "990원",
-    "name": "[걸리버] KF94 휴그린 미세먼지 방역마스크 (5매)",
-    "time": "13:00",
-    "link": "https://smartstore.naver.com/soommask/products/4828127993"
-}]
